requirements.txt
telebot==0.0.5
python-dotenv==1.0.1
discord>1.6.0